feat: player repository를 생성한다
