--- src/main/java/org/timinggame/api/room/repository/PlayerRepository.java
@@ -1,6 +1,0 @@
-package org.timinggame.api.room.repository;
-
-public interface PlayerRepository {
-
-	int getPlayerCount(final Long roomId);
-}
