+++ src/test/java/org/timinggame/api/room/RoomServiceUnitTest.java
@@ -3,7 +3,7 @@ package org.timinggame.api.room;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.timinggame.api.ServiceUnitTest;
-import org.timinggame.api.room.repository.PlayerRepository;
+import org.timinggame.api.player.repository.PlayerRepository;
 import org.timinggame.api.room.repository.RoomRepository;
 import org.timinggame.api.room.service.impl.RoomServiceV1;
